# src/main/resources/openapi/notification-api.yaml
openapi: 3.0.3
info:
  title: Group 8 Communication and Feedback API
  version: 1.0.0
  description: |
    Portable contract for in-app notifications, announcements, feedback and engagement.
    The running service also generates documentation from controllers at /v3/api-docs
    and /v3/api-docs.yaml, with interactive documentation at /swagger-ui.html.
    User operations require a Group 5 RS256 token; trusted triggers use X-Service-Key.
    Notification creation stores an inbox record, not an email or SMS.
servers:
  - url: /
paths:
  /api/notifications:
    get:
      summary: List notifications for the authenticated user
      operationId: listMyNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: A page of notifications ordered newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/{id}/read:
    patch:
      summary: Mark one notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: Updated notification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The authenticated user does not own the notification.
        '404':
          description: Notification was not found.
  /api/notifications/trigger:
    post:
      summary: Create a notification from a trusted service workflow
      description: |
        New idempotency keys create a record. Repeated keys return the original record
        even when the new payload differs. Use one distinct key per recipient and event.
        Directory verification requires a configured URL and credential. Missing settings
        or Group 5 401/403 responses allow the trusted caller to create an unverified
        recipient notification and log a warning; other provider failures return 503.
      operationId: triggerNotification
      security:
        - serviceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationTriggerRequest'
            examples:
              eventCancelled:
                value:
                  recipientId: usr-student-001
                  type: EVENT_CANCELLED
                  message: Your event has been cancelled.
                  relatedType: EVENT
                  relatedId: event-123
                  sourceService: event-service
                  idempotencyKey: event-123-cancelled-usr-student-001
      responses:
        '201':
          description: Notification created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '200':
          description: The idempotency key was already processed; the existing notification is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          description: Request validation failed.
        '401':
          description: Missing or invalid service key.
        '404':
          description: Recipient does not exist in the Group 5 directory.
        '503':
          description: Group 5 recipient directory is unavailable or returned invalid data.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: apiKey
      in: header
      name: X-Service-Key
  parameters:
    NotificationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
  schemas:
    NotificationTriggerRequest:
      type: object
      required:
        - recipientId
        - type
        - message
        - relatedType
        - sourceService
        - idempotencyKey
      properties:
        recipientId:
          type: string
          minLength: 1
          maxLength: 64
          example: usr-student-001
        type:
          $ref: '#/components/schemas/NotificationType'
        message:
          type: string
          minLength: 1
          maxLength: 2000
        relatedType:
          $ref: '#/components/schemas/RelatedType'
        relatedId:
          type: string
          maxLength: 128
          nullable: true
          description: External string id; required unless relatedType is EXTERNAL.
        sourceService:
          type: string
          minLength: 1
          maxLength: 64
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 200
    Notification:
      type: object
      required:
        - id
        - recipientId
        - type
        - message
        - relatedType
        - sourceService
        - idempotencyKey
        - isRead
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        recipientId:
          type: string
          maxLength: 64
        type:
          $ref: '#/components/schemas/NotificationType'
        message:
          type: string
        relatedType:
          $ref: '#/components/schemas/RelatedType'
        relatedId:
          type: string
          maxLength: 128
          nullable: true
        sourceService:
          type: string
        idempotencyKey:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationPage:
      type: object
      required:
        - content
        - number
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    RelatedType:
      type: string
      enum:
        - EVENT
        - REGISTRATION
        - ANNOUNCEMENT
        - RESERVATION
        - SERVICE_REQUEST
        - EXTERNAL
    NotificationType:
      type: string
      enum:
        - REGISTRATION_CONFIRMED
        - REGISTRATION_CANCELLED
        - EVENT_CANCELLED
        - EVENT_UPDATED
        - RESERVATION_STATUS
        - SERVICE_REQUEST_STATUS
...